Reduce the amount of unnecessary conversions in the Controlled gate

+++ src/main/scala/scotty/math/MathUtils.scala
@@ -11,9 +11,9 @@ object MathUtils {
     Math.rint(d * precision) / precision
   }
 
-  def toBinary(n: Long): Seq[Long] = {
+  def toBinary(n: Int): Seq[Int] = {
     @tailrec
-    def binary(acc: Seq[Long], n: Long): Seq[Long] = n match {
+    def binary(acc: Seq[Int], n: Int): Seq[Int] = n match {
       case 0 | 1 => n +: acc
       case _ => binary((n % 2) +: acc, n / 2)
     }
@@ -21,9 +21,9 @@ object MathUtils {
     binary(Seq(), n)
   }
 
-  def toBinaryPadded(n: Long, qubitCount: Int): List[Long] = {
+  def toBinaryPadded(n: Int, qubitCount: Int): List[Int] = {
     val bits = toBinary(n)
 
-    List.fill(qubitCount - bits.length)(0L) ++ bits
+    List.fill(qubitCount - bits.length)(0) ++ bits
   }
 }
